pyproject.toml
[project]
name = "ts-llm"
version = "0.5.0"
description = "Dynamical language substrate based on nonlinear attractor dynamics."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "BoggersTheFish" }
]
dependencies = [
  "numpy>=1.24.0",
  "torch>=2.4.0",
  "torchdiffeq>=0.2.0",
  "tiktoken>=0.7.0",
  "requests>=2.28.0",
  "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "matplotlib>=3.8.0",
  "plotly>=5.18.0",
  "pillow>=10.0.0",
  "pyyaml>=6.0",
  "mypy>=1.8.0",
  "types-PyYAML",
]

[project.scripts]
ts-llm = "run_attractor_llm:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["run_attractor_llm", "train"]

[tool.setuptools.packages.find]
where = ["."]
include = ["attractor_llm*", "ts_attractor*"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
follow_imports = "skip"

